pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Project to put into practice my django and langchain skills"
authors = ["Giovanna Candido Manfredo <giovannamanfredo.dev@gmail.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django-environ = "^0.12.0"
django = "^5.1.6"
mysqlclient = "^2.2.7"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.28.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
